requirements.txt
aiogram==3.0.0
python-dotenv==1.0.0
aiohttp==3.8.5
requests==2.31.0
flask==2.3.3